Компиляция в VS-2015 DinrusDMD с помощью clang-cl. Обнаружились ошибки в файлах-включениях VS-2015. Этот commit - в целях сохранения перед попыткой использования старых включений от VS-2010

+++ DRC/DMD/backend/cgen.c
@@ -690,7 +690,7 @@ L1:
     {
         if (flags & mPSW)
         {
-            switch (value)
+            switch ((int)value)
             {   case 0:
                     c = genclrreg(c,reg);
                     if (flags & 64)
